+++ back/prisma/seed.ts
@@ -12,7 +12,7 @@ async function seedDB() {
   const categoryReact = await prisma.category.findUnique({
     where: { name: 'React' },
   })
-
+  if (!categoryReact) throw new Error('Category React not found');
   const categoryNode = await prisma.category.findUnique({
     where: { name: 'Node' },
   })
@@ -25,29 +25,26 @@ async function seedDB() {
     data: [
       {
         ...users[0],
-        specializationId: categoryReact!.id,
       },
       {
         ...users[1],
-        specializationId: categoryReact!.id,
       },
       {
         ...users[2],
-        specializationId: categoryNode!.id,
       },
     ],
   })
 
-  const userAdmin = await prisma.user.findUnique({
-    where: { email: 'admin@admin.com' },
+  const userAdmin = await prisma.user.findFirst({
+    where: { name: 'Kevin Mamaqi' },
   })
 
-  const userMentor = await prisma.user.findUnique({
-    where: { email: 'mentor@mentor.com' },
+  const userMentor = await prisma.user.findFirst({
+    where: { name: 'Linux Mint' },
   })
 
-  const userRegistered = await prisma.user.findUnique({
-    where: { email: 'registered@registered.com' },
+  const userRegistered = await prisma.user.findFirst({
+    where: { name: 'Django Unchained' },
   })
 
   const topicCategories = [