+++ back/src/controllers/auth/register.ts
@@ -1,40 +1,35 @@
-import { Middleware, Context } from 'koa'
-import { prisma } from '../../prisma/client'
-import { NotFoundError } from '../../helpers/errors'
-import { UserRegister } from '../../schemas/users/userRegisterSchema'
+import { Context, Middleware } from 'koa'
 import { processMedia } from '../../helpers/processMedia'
+import { prisma } from '../../prisma/client'
+import { UserRegister } from '../../schemas/users/userRegisterSchema'
+import { ssoHandler } from '../../helpers'
 
 export const registerController: Middleware = async (ctx: Context) => {
-  const { dni, password, name, email, specialization }: UserRegister =
-    ctx.request.body
+  const {
+    dni,
+    password,
+    name,
+    email,
+    confirmPassword,
+    itineraryId,
+  }: UserRegister = ctx.request.body
 
   const media = ctx.file
 
-  const existingCategory = await prisma.category.findUnique({
-    where: { id: specialization },
+  const { id } = await ssoHandler.register({
+    dni,
+    password,
+    confirmPassword,
+    email,
+    itineraryId,
   })
-
-  if (!existingCategory) {
-    throw new NotFoundError('Category not found')
-  }
 
   const user = await prisma.user.create({
     data: {
-      dni: dni.toUpperCase(),
-      password,
+      id,
       name,
-      email,
-      specializationId: existingCategory.id,
     },
   })
-
-  if (!user || user.dni !== dni.toUpperCase()) {
-    ctx.status = 500
-    ctx.body = {
-      error: 'Database error',
-    }
-    return
-  }
 
   if (media) {
     const { mediaId } = await processMedia(media, user.id)