+++ back/src/controllers/resources/getResourcesByTopic.ts
@@ -68,7 +68,6 @@ export const getResourcesByTopicSlug: Middleware = async (ctx: Koa.Context) => {
       user: {
         select: {
           name: true,
-          email: true,
         },
       },
       vote: { select: { vote: true } },