+++ back/src/schemas/sso/ssoGetUser.ts
@@ -1,0 +1,12 @@
+import { userSchema } from '..'
+import { z } from '../../openapi/zod'
+import { ssoLoginResponseSchema } from './ssoLogin'
+
+const ssoGetUserRequestSchema = ssoLoginResponseSchema.pick({
+  authToken: true,
+})
+export type TSsoGetUserRequest = z.infer<typeof ssoGetUserRequestSchema>
+
+const ssoGetUserResponseSchema = userSchema.pick({ dni: true, email: true })
+
+export type TSsoGetUserResponse = z.infer<typeof ssoGetUserResponseSchema>