+++ front/src/__tests__/organisms/Register.test.tsx
@@ -52,8 +52,8 @@ describe('Register', () => {
       'password123AAAA'
     )
     await userEvent.selectOptions(
-      screen.getByLabelText('specialization'),
-      'React'
+      screen.getByLabelText('itineraryId'),
+      'Frontend Angular'
     )
     await userEvent.click(screen.getByTestId('accept'))
 
@@ -86,7 +86,7 @@ describe('Register', () => {
         screen.getByText(/La contrasenya ha de coincidir/i)
       ).toBeInTheDocument()
       expect(
-        screen.getByText(/Es necesario aceptar los términos legales/i)
+        screen.getByText(/És necessari acceptar els termes legals/i)
       ).toBeInTheDocument()
       expect(
         screen.getByText(