+++ packages/ui/src/components/atoms/index.ts
@@ -1,0 +1,7 @@
+import Input from './Input'
+import Label from './Label'
+import { Icon } from './Icon'
+import { Button } from './Button'
+import HamburgerMenu from './HamburgerMenu'
+
+export { Input, Label, Button, HamburgerMenu, Icon }