+++ sso/src/controllers/token/validate.ts
@@ -1,20 +1,13 @@
 import { Context, Middleware } from 'koa'
 import { InvalidCredentials } from '../../utils/errors'
-import { verifyToken } from '../../utils/auth'
-import { ValidateSchema } from '../../schemas/token/validateSchema'
 
 export const validate: Middleware = async (ctx: Context) => {
-  const { authToken } = ctx.request.body as ValidateSchema
+  const { id } = ctx
 
-  if (!authToken) {
-    throw new InvalidCredentials()
-  }
-
-  const validToken = verifyToken(authToken)
-
-  if (!validToken) {
+  if (!id) {
     throw new InvalidCredentials()
   } else {
-    ctx.status = 204
+    ctx.status = 200
+    ctx.body = { id }
   }
 }