+++ sso/src/schemas/token/tokenSchema.ts
@@ -1,6 +1,8 @@
 import { z } from '../../openapi/zod'
 
-export const tokenSchema = z.object({
-  authToken: z.string().nonempty(),
-  refreshToken: z.string().nonempty(),
-})
+export const tokenSchema = z
+  .object({
+    authToken: z.string().nonempty(),
+    refreshToken: z.string().nonempty(),
+  })
+  .extend({ id: z.string().cuid2() })